Move some templates and jobs to jenkins-jobs
This allows MCI to re-use them instead of doing code-copy.
