+++ jenkins-jobs/multijob_phase.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 #
-# Copyright (C) 2016 Harald Sitter <sitter@kde.org>
+# Copyright (C) 2015-2016 Harald Sitter <sitter@kde.org>
 #
 # This library is free software; you can redistribute it and/or
 # modify it under the terms of the GNU Lesser General Public
@@ -18,15 +18,23 @@
 # You should have received a copy of the GNU Lesser General Public
 # License along with this library.  If not, see <http://www.gnu.org/licenses/>.
 
-require_relative '../template'
+require_relative 'template'
 
 # a phase partial
-class PackagingSCMTemplate < Template
-  attr_reader :scm
+class MultiJobPhase < Template
+  # @!attribute [r] phase_name
+  #   @return [String] name of the phase
+  attr_reader :phase_name
 
-  # @param scm SCM instance to templatify
-  def initialize(scm:)
-    super("packaging-scms/#{scm.type}.xml.erb")
-    @scm = scm
+  # @!attribute [r] phased_jobs
+  #   @return [Array<String>] name of the phased jobs
+  attr_reader :phased_jobs
+
+  # @param phase_name see {#phase_name}
+  # @param phased_jobs see {#phased_jobs}
+  def initialize(phase_name:, phased_jobs:)
+    super("#{File.basename(__FILE__, '.rb')}.xml.erb")
+    @phase_name = phase_name
+    @phased_jobs = phased_jobs
   end
 end
