+++ jenkins-jobs/nci/project.rb
@@ -24,7 +24,7 @@ require_relative '../binarier'
 require_relative '../publisher'
 require_relative 'lintcmakejob'
 require_relative 'lintqmljob'
-require_relative 'multijob_phase'
+require_relative '../multijob_phase'
 
 # Magic builder to create an array of build steps
 class ProjectJob < JenkinsJob
