+++ jenkins-jobs/templates/multijob_phase.xml.erb
