+++ jenkins-jobs/templates/packaging-scms/bzr.xml.erb
