+++ jenkins-jobs/templates/packaging-scms/git.xml.erb
