+++ jenkins-jobs/templates/project.xml.erb
